// web/src/models/comic.models.ts
export interface Comic {
    month: string;
    num: number;
    link: string;
    year: string;
    news: string;
    safe_title: string;
    transcript: string;
    alt: string;
    img: string;
    title: string;
    day: string;
}

export interface ComicWrapper {
    comic: Comic;
    viewCounts: number;
}